+++ tests/Order/OrderServiceTest.php
@@ -15,6 +15,7 @@ use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\OrderNotFound;
 use Wizaplace\SDK\Order\AfterSalesServiceRequest;
 use Wizaplace\SDK\Order\CreateOrderReturn;
+use Wizaplace\SDK\Order\OrderAdjustment;
 use Wizaplace\SDK\Order\OrderCommitmentCommand;
 use Wizaplace\SDK\Order\OrderReturnStatus;
 use Wizaplace\SDK\Order\OrderService;
@@ -283,6 +284,23 @@ final class OrderServiceTest extends ApiTestCase
     {
         $this->buildOrderService('admin@wizaplace.com', 'password')->createOrderAdjustment(10, 3230927120, 10);
         static::assertSame(9.01, $this->buildOrderService()->getOrder(10)->getTotal());
+    }
+
+    public function testGetAdjustments(): void
+    {
+        $orderService = $this->buildOrderService('admin@wizaplace.com', 'password');
+        $adjustments = $orderService->getAdjustments(1);
+        /** @var OrderAdjustment $adjustment */
+        $adjustment = current($adjustments);
+        static::assertInternalType('array', $adjustments);
+        static::assertCount(1, $adjustments);
+        static::assertSame(2085640488, $adjustment->getItemId());
+        static::assertSame(70.9, $adjustment->getOldPrice());
+        static::assertSame(67.9, $adjustment->getNewPrice());
+        static::assertSame(70.9, $adjustment->getOldTotal());
+        static::assertSame(67.9, $adjustment->getNewTotal());
+        static::assertSame(7, $adjustment->getCreatedBy());
+        static::assertInstanceOf(\DateTime::class, $adjustment->getCreatedAt());
     }
 
     private function buildOrderService(string $email = 'customer-1@world-company.com', $password = 'password-customer-1'): OrderService